change to Integer

+++ springbootWithPostgresql/src/main/java/com/example/springbootWithPostgresql/repository/ProdRepo.java
@@ -4,5 +4,5 @@ package com.example.springbootWithPostgresql.repository;
 import com.example.springbootWithPostgresql.entity.ProductEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdRepo extends JpaRepository<ProductEntity, Long> {
+public interface ProdRepo extends JpaRepository<ProductEntity, Integer> {
 }
